+++ reusables/math/PlatonicSolid.h
@@ -72,8 +72,8 @@ namespace math
     PlatonicSolid& operator=(const PlatonicSolid&);
     virtual ~PlatonicSolid();
   
-    virtual int getNumberOfPolygons() const { return mPolygons.size(); }
-    virtual int getNumberOfVertices() const { return mVertex.size(); } 
+    virtual int getNumberOfPolygons() const { return (int)mPolygons.size(); }
+    virtual int getNumberOfVertices() const { return (int)mVertex.size(); } 
     virtual const Polygon& getPolygon( int ) const;
     virtual const vector<Polygon>& getPolygons() const { return mPolygons; } 
     virtual const type getType() const {return mType;}
